feat: 🎸 Check phone number
Added a function to check if the sent phone number is stored in the
database

+++ telephone-address-book-api/Program.cs
@@ -1,4 +1,6 @@
+using Microsoft.AspNetCore.Mvc;
 using MySql.Data.MySqlClient;
+using System.Threading.Tasks;
 using System.Data;
 using dotenv.net;
 using Dapper;
@@ -28,15 +30,12 @@ app.MapGet("/book", async () =>
     try
     {
         conn.Open();
-        var bookList = await conn.QueryAsync("Select * FROM book;");
+        var data = await conn.QueryAsync("Select * FROM book;");
         conn.Close();
-        return bookList.ToList();
+
+        return Results.Json(new { status = 200, data = data.ToList(), message = "Data Retrieved" });
     }
-    catch (System.Exception e)
-    {
-        Console.WriteLine(e.Message);
-        throw;
-    }
+    catch (System.Exception e) { return Results.Json(new { status = 400, error = e.Message }); }
 });
 
 app.MapGet("/book/{phone}", async (int phone) =>
@@ -46,15 +45,14 @@ app.MapGet("/book/{phone}", async (int phone) =>
     try
     {
         conn.Open();
-        var result = await conn.QueryAsync("Select * FROM book WHERE phone='" + phone + "';");
+        var data = await conn.QueryAsync("Select * FROM book WHERE phone='" + phone + "';");
         conn.Close();
-        return result;
+
+        if (!data.Any()) { return Results.Json(new { status = 404, data = "This number doesn't exist in the database" }); };
+
+        return Results.Json(new { status = 200, data = data.First(), message = "Data Retrieved" });
     }
-    catch (System.Exception e)
-    {
-        Console.WriteLine(e.Message);
-        throw;
-    }
+    catch (System.Exception e) { return Results.Json(new { status = 400, error = e.Message }); }
 });
 
 app.MapPost("/book", async (Book book) =>
@@ -65,17 +63,38 @@ app.MapPost("/book", async (Book book) =>
     {
         conn.Open();
 
-        await conn.QueryAsync("INSERT INTO book (phone, first_name, last_name, household, email) " + "VALUES (@Phone, @First_name, @Last_name, @Household, @Email)", new { Phone = book.Phone, First_name = book.First_name, Last_name = book.Last_name, Household = book.Household, Email = book.Email });
+        var data = new
+        {
+            Phone = book.Phone,
+            First_name = book.First_name,
+            Last_name = book.Last_name,
+            Household = book.Household,
+            Email = book.Email
+        };
+
+        await conn.QueryAsync("INSERT INTO book (phone, first_name, last_name, household, email) " + "VALUES (@Phone, @First_name, @Last_name, @Household, @Email)", data);
 
         conn.Close();
-        return "Added '" + book.Phone + "' to book";
+
+        return Results.Json(new { status = 201, data, message = "Data Created" });
     }
-    catch (System.Exception e)
-    {
-        Console.WriteLine(e.Message);
-        throw;
-    }
+    catch (System.Exception e) { return Results.Json(new { status = 400, error = e.Message }); }
 });
+
+async Task<bool> IsNumberRegisteredAsync(int phone)
+{
+    try
+    {
+        var conn = ConnectDB();
+        var data = await conn.QueryAsync("SELECT * FROM book WHERE phone = @Phone;", new { Phone = phone });
+        conn.Close();
+
+        if (!data.Any()) { return false; }
+
+        return true;
+    }
+    catch (System.Exception) { return false; }
+}
 
 app.MapPut("/book/{phone}", async (int phone, Book book) =>
 {
@@ -83,46 +102,36 @@ app.MapPut("/book/{phone}", async (int phone, Book book) =>
 
     try
     {
+        if (!await IsNumberRegisteredAsync(phone)) { return Results.Json(new { status = 404, data = "This number doesn't exist in the database" }); }
+
         conn.Open();
 
-        string updateQuery = "UPDATE book SET ";
-        bool isFirstField = true;
+        var data = new
+        {
+            Phone = phone,
+            NewPhone = book.Phone == 0 ? phone : book.Phone,
+            FirstName = book.First_name,
+            LastName = book.Last_name,
+            Household = book.Household,
+            Email = book.Email
+        };
 
-        if (book.First_name != null)
-        {
-            updateQuery += "first_name = @FirstName";
-            isFirstField = false;
-        }
-        if (book.Last_name != null)
-        {
-            updateQuery += isFirstField ? "" : ", ";
-            updateQuery += "last_name = @LastName";
-            isFirstField = false;
-        }
-        if (book.Household != null)
-        {
-            updateQuery += isFirstField ? "" : ", ";
-            updateQuery += "household = @Household";
-            isFirstField = false;
-        }
-        if (book.Email != null)
-        {
-            updateQuery += isFirstField ? "" : ", ";
-            updateQuery += "email = @Email";
-        }
+        string query = "UPDATE book SET phone = @NewPhone";
 
-        updateQuery += " WHERE phone = @Phone";
+        if (book.First_name != null) { query += ", first_name = @FirstName"; }
+        if (book.Last_name != null) { query += ", last_name = @LastName"; }
+        if (book.Household != null) { query += ", household = @Household"; }
+        if (book.Email != null) { query += ", email = @Email"; }
 
-        await conn.QueryAsync(updateQuery, new { Phone = phone, FirstName = book.First_name, LastName = book.Last_name, Household = book.Household, Email = book.Email });
+        query += " WHERE phone = @Phone";
+
+        await conn.QueryAsync(query, data);
 
         conn.Close();
-        return "'" + book.Phone + "' updated";
+
+        return Results.Json(new { status = 201, data, message = "Data Updated" });
     }
-    catch (System.Exception e)
-    {
-        Console.WriteLine(e.Message);
-        throw;
-    }
+    catch (System.Exception e) { return Results.Json(new { status = 400, error = e.Message }); }
 });
 
 app.MapDelete("/book/{phone}", async (int phone) =>
@@ -131,16 +140,15 @@ app.MapDelete("/book/{phone}", async (int phone) =>
 
     try
     {
+        if (!await IsNumberRegisteredAsync(phone)) { return Results.Json(new { status = 404, data = "This number doesn't exist in the database" }); }
+
         conn.Open();
         await conn.QueryAsync("Delete FROM book WHERE phone='" + phone + "';");
         conn.Close();
-        return "Regist deleted '" + phone + "'";
+
+        return Results.Json(new { status = 201, message = $"{phone} data was Deleted" });
     }
-    catch (System.Exception e)
-    {
-        Console.WriteLine(e.Message);
-        throw;
-    }
+    catch (System.Exception e) { return Results.Json(new { status = 400, error = e.Message }); }
 });
 
 app.Run();
